+++ kit/dapp/test/scripts/demo/data/demo-country-codes.ts
@@ -1,0 +1,2 @@
+export const DE_COUNTRY_CODE = "276";
+export const JP_COUNTRY_CODE = "392";
