+++ kit/dapp/test/scripts/demo/data/demo-users.ts
@@ -4,7 +4,6 @@ import {
   DEFAULT_ISSUER,
   User,
 } from "@test/fixtures/user";
-import { from } from "dnum";
 
 export const ADMIN: User = {
   ...DEFAULT_ADMIN,
@@ -38,29 +37,3 @@ export const JAPANESE_INVESTOR: User = {
   firstName: "Japanese",
   lastName: "Investor",
 };
-
-export const DE_COUNTRY_CODE = "276";
-export const JP_COUNTRY_CODE = "392";
-
-export const BONDS = [
-  {
-    name: "Bund7",
-    symbol: "BUND7",
-    isin: "DE000BU27014",
-    yieldRate: 250,
-    cap: from(8_000_000_000, 18),
-    issueDate: new Date(2026, 8, 24),
-    maturityDate: new Date(2027, 8, 24),
-    countries: [DE_COUNTRY_CODE],
-  },
-  {
-    name: "Bund10",
-    symbol: "BUND10",
-    isin: "DE000BU2Z056",
-    yieldRate: 260,
-    cap: from(26_000_000_000, 18),
-    issueDate: new Date(2026, 8, 24),
-    maturityDate: new Date(2027, 8, 24),
-    countries: [DE_COUNTRY_CODE, JP_COUNTRY_CODE],
-  },
-];
